Re-organize `libertem_holo.base.mask`
Move contents to `libertem_holo.base.filters`; remove the old line
filter implementation.
